added warning and rename a file
