+++ bin/help.js
@@ -1,5 +1,5 @@
-import formatting from "../src/formatting.mjs";
-import l10n from "../localization/getstring.mjs";
+import formatting from "../src/formatting.js";
+import l10n from "../localization/getstring.js";
 import osLocale from "os-locale";
 const locale = osLocale.sync();
 var title = l10n("1host.js Help", locale);
