+++ localization/getstring.js
